Simplify the implementation, put everything in the same file
