--- internal/iopool/buffer.go
@@ -1,11 +1,0 @@
-package iopool
-
-type buffer struct {
-	b []byte
-}
-
-func newBuffer(len int) *buffer {
-	return &buffer{
-		b: make([]byte, len),
-	}
-}
