+++ internal/iopool/copy_test.go
@@ -2,7 +2,6 @@ package iopool
 
 import (
 	"bytes"
-	"io"
 	"math/rand"
 	"sync"
 	"testing"
@@ -16,9 +15,6 @@ func TestBuffers(t *testing.T) {
 		wg.Add(1)
 		go func() {
 			for i := 0; i < 100; i++ {
-				buf := get()
-				assert.True(t, len(buf.b) >= _defaultCapacity, "Expected large buffer")
-
 				inputBytes := make([]byte, rand.Intn(5000)+20)
 				_, err := rand.Read(inputBytes)
 				assert.NoError(t, err, "Unexpected error from rand.Read")
@@ -27,11 +23,7 @@ func TestBuffers(t *testing.T) {
 				outputBytes := make([]byte, 0, len(inputBytes))
 				writer := bytes.NewBuffer(outputBytes)
 
-				io.CopyBuffer(writer, reader, buf.b)
-
-				assert.Equal(t, outputBytes[:20], inputBytes[:20], "Input and output from copy did not match")
-
-				put(buf)
+				Copy(writer, reader)
 			}
 			wg.Done()
 		}()
